Reorganized main tests
Moved `MiniExcelOpenXmlTests` and `MiniExcelOpenXmlAsyncTests` classes to their own folder and separated models into their own file
